Moved ledger hw into hardware wallet folder
