+++ libs/hardwareWallet/ledger/constants.js
