+++ libs/hardwareWallet/ledger/ledgerDeviceListener/subscribeToLedgerDevices.js
@@ -1,4 +1,4 @@
-import { LEDGER_HW_HID_EVENT } from '@libs/ledgerHardwareWallet/constants';
+import { LEDGER_HW_HID_EVENT } from '@libs/hardwareWallet/ledger/constants';
 
 const IPC = window.ipc;
