+++ libs/hardwareWallet/ledger/ledgerLiskAppIPCChannel/serverLedgerHWCommunication/utils.js
