--- libs/ledgerHardwareWallet/initLedgerHardwareWalletIPC.js
@@ -1,7 +1,0 @@
-import { ledgerDeviceListener } from '@libs/ledgerHardwareWallet/ledgerDeviceListener';
-import { ledgerLiskAppIPCChannel } from '@libs/ledgerHardwareWallet/ledgerLiskAppIPCChannel';
-
-export function initLedgerHardwareWalletIPC(win) {
-  ledgerDeviceListener(win);
-  ledgerLiskAppIPCChannel();
-}
