+++ src/modules/transaction/configuration/statusConfig.js
@@ -1,7 +1,7 @@
 /* eslint-disable complexity */
 import { isEmpty } from 'src/utils/helpers';
 import { txStatusTypes } from '@transaction/configuration/txStatus';
-import { LEDGER_HW_IPC_CHANNELS } from '@libs/ledgerHardwareWallet/constants';
+import { LEDGER_HW_IPC_CHANNELS } from '@libs/hardwareWallet/ledger/constants';
 import { transactionToJSON, getNumberOfSignatures, joinModuleAndCommand } from '../utils';
 import { MODULE_COMMANDS_NAME_MAP } from './moduleCommand';
